+++ beacon-chain/node/BUILD.bazel
@@ -7,6 +7,7 @@ go_library(
     visibility = ["//beacon-chain:__subpackages__"],
     deps = [
         "//shared:go_default_library",
+        "//shared/debug:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@com_github_urfave_cli//:go_default_library",
     ],
