+++ beacon-chain/node/node.go
@@ -7,6 +7,7 @@ import (
 	"syscall"
 
 	"github.com/prysmaticlabs/geth-sharding/shared"
+	"github.com/prysmaticlabs/geth-sharding/shared/debug"
 	log "github.com/sirupsen/logrus"
 	"github.com/urfave/cli"
 )
@@ -56,6 +57,7 @@ func (b *BeaconNode) Start() {
 				log.Info("Already shutting down, interrupt more to panic.", "times", i-1)
 			}
 		}
+		debug.Exit() // Ensure trace and CPU profile data are flushed.
 		panic("Panic closing the beacon node")
 	}()
