+++ sharding/BUILD.bazel
@@ -8,6 +8,7 @@ go_library(
     deps = [
         "//sharding/node:go_default_library",
         "//sharding/utils:go_default_library",
+        "//shared/debug:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@com_github_urfave_cli//:go_default_library",
     ],
