+++ sharding/node/BUILD.bazel
@@ -19,6 +19,7 @@ go_library(
         "//sharding/types:go_default_library",
         "//sharding/utils:go_default_library",
         "//shared:go_default_library",
+        "//shared/debug:go_default_library",
         "@com_github_ethereum_go_ethereum//event:go_default_library",
         "@com_github_ethereum_go_ethereum//node:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
