+++ sharding/utils/BUILD.bazel
@@ -4,7 +4,6 @@ go_library(
     name = "go_default_library",
     srcs = [
         "customflags.go",
-        "debug.go",
         "flags.go",
         "service.go",
     ],
@@ -13,7 +12,6 @@ go_library(
     deps = [
         "//sharding/params:go_default_library",
         "@com_github_ethereum_go_ethereum//node:go_default_library",
-        "@com_github_fjl_memsize//memsizeui:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@com_github_urfave_cli//:go_default_library",
     ],
