+++ shared/debug/BUILD.bazel
@@ -1,0 +1,13 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["debug.go"],
+    importpath = "github.com/prysmaticlabs/geth-sharding/shared/debug",
+    visibility = ["//visibility:public"],
+    deps = [
+        "@com_github_fjl_memsize//memsizeui:go_default_library",
+        "@com_github_sirupsen_logrus//:go_default_library",
+        "@com_github_urfave_cli//:go_default_library",
+    ],
+)
